requirements.txt
firebase_admin==6.5.0
mysql_connector_repackaged==0.3.1
numpy==1.26.4
opencv_python==4.9.0.80
Pillow==10.3.0
reportlab==4.1.0
Requests==2.31.0
streamlit==1.33.0
streamlit_authenticator==0.1.5
ultralytics==8.2.1
